requirements.txt
aniso8601==9.0.1
antiorm==1.2.1
click==8.1.3
colorama==0.4.4
db==0.1.1
db-sqlite3==0.0.1
decorator==5.1.1
Flask==2.1.2
Flask-RESTful==0.3.9
greenlet==1.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pbr==5.9.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.37
sqlalchemy-migrate==0.13.0
sqlparse==0.4.2
Tempita==0.5.2
Werkzeug==2.1.2
zipp==3.8.0
